Clean paths and more in style.css

+++ demo/src/main/java/com/example/ClotheDAO.java
@@ -10,6 +10,6 @@ import lombok.AllArgsConstructor;
 public class ClotheDAO {
     private String name; 
     private BigDecimal price;
-    private byte quantity;
+    private short quantity;
     private String description;
 }
